Proxy/requirements.txt
fastapi
uvicorn[standard]
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
